Finished the Game.
The final update. 
Includes two new upgrades to balance the game, and added the final ending and secret ending extra bit. Also fixed some exploits and glitches.

There may be another small one in the feature for bug-fixes and the lore update if I get around to it, but it's not 100% happening.